feat: changed modal to touchable without feedback
